Refactored suggested commmunities components
