+++ packages/commonwealth/client/scripts/views/pages/user_dashboard/TrendingCommunitiesPreview/CommunityPreviewCard/index.ts
@@ -1,0 +1,1 @@
+export { CommunityPreviewCard } from './CommunityPreviewCard';
