+++ packages/commonwealth/client/scripts/views/pages/user_dashboard/TrendingCommunitiesPreview/TrendingCommunitiesPreview.scss
@@ -1,0 +1,32 @@
+@import '../../../../../styles/shared';
+
+.TrendingCommunitiesPreview {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  position: fixed;
+  right: 50px;
+  float: right;
+  top: 100px;
+  z-index: 1;
+  flex-grow: 1;
+  overflow-y: auto;
+  max-height: 100vh;
+  padding-bottom: 100px;
+
+  @include mediumSmallInclusive {
+    display: none;
+  }
+
+  .community-preview-cards-collection {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+  }
+
+  .buttons {
+    display: flex;
+    gap: 16px;
+    justify-content: center;
+  }
+}
