+++ packages/commonwealth/client/scripts/views/pages/user_dashboard/TrendingCommunitiesPreview/index.ts
@@ -1,0 +1,1 @@
+export { TrendingCommunitiesPreview } from './TrendingCommunitiesPreview';
