+++ packages/commonwealth/client/styles/pages/user_dashboard/index.scss
@@ -8,7 +8,7 @@
   z-index: 0;
   @media (min-width: $breakpoint-medium-min-px) {
     /* CommunityPreviewCard 392px width +
-       DashboardCommunitiesPreview 50px right edge */
+       TrendingCommunitiesPreview 50px right edge */
     margin-right: 442px;
   }
